refactor(ale_claw): Tier 4b — group tool modules into tools/ subpackage
Moves the 7 tool-cluster modules (tools, tools_fs, tools_shell, tools_web,
_tool_utils, _paths, fs_backends) into harness/tools/ via git mv (names kept).
Sibling imports stay single-dot; non-sibling root imports become double-dot
(tools.py: ..analyze_image/..memory/..milestone/..subagent.*; tools_fs.py:
..image_sanitization). External importers (__init__, deployer's build_tools, and
the fs/exec/tools tests) updated to .tools.X / harness.tools.X. @register_tool
side effects preserved via the updated __init__ import. Pure move; build_tools
imports clean; full suite green (928 passed, 3 skipped).
