+++ ale_run/agents/ale_claw/harness/__init__.py
@@ -95,10 +95,10 @@ from .subagent.subagent_registry import (
 )
 from .subagent.subagent_tools import DelegateGeneralTool, DelegateGUITool, SubagentsTool
 from .thinking import ThinkingConfig, ThinkLevel, resolve_thinking_default
-from .tools import ToolLoggingCallback, build_tools, get_tool_summaries
-from .tools_fs import EditFileTool, ReadFileTool, WriteFileTool
-from .tools_shell import ExecTool
-from .tools_web import WebFetchTool, WebSearchTool
+from .tools.tools import ToolLoggingCallback, build_tools, get_tool_summaries
+from .tools.tools_fs import EditFileTool, ReadFileTool, WriteFileTool
+from .tools.tools_shell import ExecTool
+from .tools.tools_web import WebFetchTool, WebSearchTool
 from .transcript import group_step_output
 
 # Side-effect import — registers the OpenRouter unified loop with
