+++ ale_run/agents/ale_claw/harness/tools/tools_shell.py
