+++ tests/ale_claw/test_exec_tool.py
@@ -16,7 +16,7 @@ from dataclasses import dataclass
 from unittest.mock import AsyncMock, MagicMock
 
 from agent.tools.base import TOOL_REGISTRY
-from ale_run.agents.ale_claw.harness.tools_shell import (
+from ale_run.agents.ale_claw.harness.tools.tools_shell import (
     ExecTool,
     _DEFAULT_MAX_OUTPUT_CHARS,
     _DEFAULT_TIMEOUT_SECONDS,
@@ -94,7 +94,7 @@ class TestRegistration:
     def test_exec_in_build_tools_default_list(self):
         from unittest.mock import patch
 
-        from ale_run.agents.ale_claw.harness.tools import build_tools
+        from ale_run.agents.ale_claw.harness.tools.tools import build_tools
 
         session = MagicMock()
         session.interface = MagicMock()
