+++ tests/ale_claw/test_fs_backends.py
@@ -19,13 +19,13 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from ale_run.agents.ale_claw.harness.fs_backends import (
+from ale_run.agents.ale_claw.harness.tools.fs_backends import (
     FilesystemRegistry,
     HostBackend,
     VMBackend,
     detect_host_workspace_root,
 )
-from ale_run.agents.ale_claw.harness.tools_fs import (
+from ale_run.agents.ale_claw.harness.tools.tools_fs import (
     EditFileTool,
     ReadFileTool,
     WriteFileTool,
